customer server actions validation

+++ actions/categories/delete-category.ts
@@ -1,9 +1,16 @@
 "use server"
 
+import { auth } from "@/lib/auth/auth"
 import prisma from "@/lib/db/db"
 import { revalidatePath } from "next/cache"
 
 export async function deleteCategory({ id }: { id: string }) {
+  const session = await auth()
+
+  if (session?.user.role !== "ADMIN") {
+    return { error: "No autorizado." }
+  }
+
   try {
     const category = await prisma.category.delete({
       where: { id },
